refactor(simulation): split simulation controller by feature
