--- src/main/java/com/renewsim/backend/simulation_service/web/controller/SimulationRequestContext.java
@@ -1,6 +1,0 @@
-package com.renewsim.backend.simulation_service.web.controller;
-
-record SimulationRequestContext(
-                String username,
-                boolean isAdmin) {
-}
